Add utils folder and various utility functions
