+++ src/utils/buildPolishNotation.js
